Added: Functionality to check True or False

+++ linecomparison/Point.java
@@ -1,5 +1,5 @@
 package linecomparison;
 
 public class Point {
-	float x, y;
+	double x, y;
 }
